# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ADP API
  version: 0.3.0
  description: >
    Native and compat-plane HTTP surface for ADP — see docs/pragmatic_mvp.md
    §2.4 for what is in scope and §2.5 for what is deliberately cut.

    Two planes share this file. `/api/v3` is the GitHub-compatible surface, kept
    faithful enough that unmodified `gh` works against it. `/api/adp` is the
    native plane, which has no GitHub analogue: workspaces, candidate sets,
    sessions and checkpoints, the operation log, and evidence bundles.

    Completeness is enforced, not aspirational: server/src/spec-coverage.test.ts
    enumerates the server's real route table and fails if this file and the
    implementation describe different APIs in either direction. Deliberate
    omissions are the git smart-HTTP wire protocol (specified by git itself and
    proxied verbatim), the loopback-only receive-path hook endpoints, and the
    inbound GitHub webhook receiver, whose payload contract is GitHub's.

    `info.version` above is the contract version, and every response carries it
    in an `ADP-API-Version` header — including 401s and 404s, so a generated
    client can pin before it authenticates. The git smart-HTTP routes are the
    exception: they are proxied verbatim and are not part of this contract.
    See docs/api-compatibility.md for what a version bump promises.
servers:
  - url: /
paths:
...
  /api/v3/repos/{owner}/{repo}/changes:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    post:
      x-required-scope: "repo:write"
      summary: >
        Record a typed change: an existing git commit plus intent and
        provenance, server-signed. No GitHub shape — native ADP metadata
        living beside the commit (docs/pragmatic_mvp.md §2.2). Recorded via
        explicit call; not yet triggered automatically by a git push.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChangeRequest"
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Repository not found
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: git_sha is not a commit in this repo, or intent_id doesn't belong to it
  /api/v3/repos/{owner}/{repo}/changes/{id}:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      security:
        - bearerAuth: []
      x-required-scope: "repo:read"
      summary: Get a signed change record
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
...
  /api/v3/repos/{owner}/{repo}/dependency-admission:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    post:
      x-required-scope: "repo:write"
      summary: >
        Admission verdict for a lockfile diff — three-way `admit`/`quarantine`/
        `block` per package, recorded as an ordinary gate result. `quarantine`
        reports gate status `pending`, which `gates_green` already treats as
        not-green, so a repo opts in by naming the gate in adp.yaml.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Verdict recorded
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation failed
...
components:
  parameters:
    owner:
      name: owner
      in: path
      required: true
      schema:
        type: string
    repo:
      name: repo
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [message]
      description: >
        #97: the one error shape every 4xx carries. `errors` appears on
        validation failures (422) and is a deliberate projection of the
        server's internal validator output — path, message, and a stable
        code — never the validator's raw issue objects, whose shape belongs
        to a library this contract must outlive.
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, message, code]
            properties:
              path:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              message:
                type: string
              code:
                type: string
...
    CreateChangeRequest:
      type: object
      required: [git_sha]
      properties:
        git_sha:
          type: string
          pattern: "^[0-9a-f]{40}$"
        intent_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
...